Axis is now an interface, AxisImpl implements it
